requirements.txt
astroid==1.6.4
dj-database-url==0.5.0
Django==2.0.5
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
flake8==3.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.4
six==1.11.0
whitenoise==3.3.1
wrapt==1.10.11
